Added representation class for data structures
The new class mainly represents what is stored in a common TCA entry.
There is a resolver for building the neccessary information out of the
plain TCA entry. Currently the data structure object is created for a
record object and injected to it while it is being added to a form
object. The form builder then uses it instead of the old TCA information
for building the object tree.
This is a preparation for easier implementation of FlexForms.

Still to be done is some proper caching mechanism for the data structure
resolver.