[BUGFIX] Pages with Enforce Protocol (SSL) may generate wrong URL
If TYPO3 is installed in a subfolder, you won't need a absRefPrefix set because normally links are relative.
If a Protocol change will (SSL -> PLAIN or vice versa) happen, the URL has to be generated absolute.
This will fail in the described behaviour.

Notes:
This one has been reported by Michael Maier.
Attached patch is my "easier to read" version doing the same as _v2 of Jigal and the result of talking through all cases which may arise.

Change-Id: Ie5821361b578b6f1fa3612ac9bfa44dd476c9c76
Resolves: #M17118
Reviewed-on: http://review.typo3.org/953
Reviewed-by: Tobias Liebig <mail_typo3.org@etobi.de>
Reviewed-by: Andreas Wolf <andreas.wolf@ikt-werk.de>
Tested-by: Andreas Wolf <andreas.wolf@ikt-werk.de>
Reviewed-by: Sebastian Fischer <typo3@evoweb.de>
Reviewed-by: Helmut Hummel <typo3@helmut-hummel.de>
Tested-by: Helmut Hummel <typo3@helmut-hummel.de>
Reviewed-by: Jigal van Hemert <jigal@xs4all.nl>