[FEATURE] Fields with FlexFormSection get proper type set
This enables the FormBuilder to recognize fields that contain FlexForm
sections the same way it recognizes the other field types. Thus we can
create a simple element class for FlexForm sections now.

This commit also refactors some unit tests and adds an additional one
that checks for correct extraction of the field config from the DS XML.