[TASK] Add widget registry and factory
This commit also introduces a new subarray of TYPO3_CONF_VARS, TCEFORMS.
This array is designated to contain all configuration around TCEforms,
currently this is most notably the widget types. Currently, the array is
only a stub, but it will be filled as soon as there are useful widgets
around. Additionally, the wizard registry will be extend it when the
corresponding registration method is called. This will help in enabling
people to create their own widgets.

The whole type concept introduced in factory/registry is somehow
borrowed from the ExtJs Xtype concept. It will enable widget creation
with just a shortcut instead of the long (and thus error-prone) class
names.