[FEATURE] Add WidgetBuilder class
The builder is responsible for all "high-level" widget building tasks,
while the factory is "only" responsible for low-level plumbing. The
factory would also be the one to keep track of single-instance entities
(not singletons), if we ever need any.

The builder will be called by the Form/Context after a record has been
added.