[TASK] (TCA) Large refactorings in Type class
This throws away a lot of old stuff, among others the whole sheet
handling and the different static constructors.

Constructing Type objects should now be done via the normal constructor
with the configuration of the type as the third parameter. For
array-based TCA, this can be directly taken from the [types] section
entry. For FlexForms, a corresponding entry has to be created inside the
FlexFormDataStructure class (tbd).

Additionally, using sheets, palettes etc. inside TCA is deprecated from
now on. Instead, these elements will be mapped to corresponding widgets
by a yet-to-write transformer for showitem strings. The result of this
transformation will be a nested array of widget configurations, as is
used by e.g. ExtJS. This can then be used for producing widget trees
with the WidgetBuilder class.