[TASK] (TCA) Convert showitem strings to widget config arrays
The new methods are able to handle showitem strings from type and
palette entries. Output is an array of widget configurations that could
be directly used with the TCEforms widget builder class.

Palettes in type showitem strings are currently inserted just as a
reference to a "block". The TCEforms widget builder is missing support
for this (see story #26322).

Still missing is a handling for blocks to re-use parts of the
configuration. This could be used for e.g. palettes, which are already
"abused" as building blocks in some part of the core configuration.

Resolves #26323