[FEATURE] WidgetBuilder: Create widget trees for records
The widget tree creation is a two-step process: First an "abstract"
widget tree for a TCA type is built. Second, this abstract tree is bound
to a concrete record.

In the first step, all fields are added as FieldProxy objects, which are
later on replaced by concrete field widgets in the second step.

A definitive must-have is caching for the abstract trees, maybe even
static caching over request boundaries. This should be added inside
buildWidgetTreeForType().