[+FEATURE] Add abstract renderer
The rendering infrastructure relies on a renderer that is instantiated
by the Forms consumer. A widget tree may be rendered beginning from
every container widget, though it is debatable whether this makes sense
in every case. The normal case should be to render from the very
beginning of the TCEforms widget tree, or a defined subtree root (e.g.
for AJAX calls).

There are no concrete renderers in this commit; they will be added later
on. The variation point for concrete renderers is mainly the
renderWidget() method.