[BUGFIX] Record: Language overlay breaks default language
The record object returned an empty string for non-localized fields or
if the record had the default lanugage (and some localization modes were
set). This is incorrect, it should return the default language's value.

This bug was visible when trying to get the type value for a record - it
yielded an incorrect result.